requirements.txt
pdfservices-sdk>=4.0.0
python-dotenv>=1.0.0
plyer>=2.1.0
pypdf>=4.0.0
